requirements.txt
Django==4.0.3
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers